Change folder structure of the project by adding layouts folder
