+++ src/App.js
@@ -1,13 +1,8 @@
-import { Routes, Route } from 'react-router-dom';
-import routes from './routes';
+import Layout from './layouts/Layout';
 
 function App() {
   return (
-    <Routes>
-      {routes.map((route) => (
-        <Route key={`key${route.path}`} path={route.path} element={route.element} exact={route.exact} />
-      ))}
-    </Routes>
+    <Layout />
   );
 }
