+++ src/pages/Homepage.js
@@ -1,28 +1,11 @@
-import React, { useState } from 'react';
-import { FaBars } from 'react-icons/fa';
-import Navbar from '../components/Navbar';
-import style from './Homepage.module.css';
+import React from 'react';
 
 export default function Homepage() {
-  const [menu, setMenu] = useState(true);
-
-  const clickHandler = () => {
-    setMenu(!menu);
-  };
-
   return (
-    <div className={style.homepage_container}>
-
-      <div className={style.navbar}>
-        { menu && <Navbar /> }
-      </div>
-      <div className={style.content}>
-        <button type="button" className={style.menutoggle} aria-label="Menu toggle" onClick={() => { clickHandler(); }}>
-          <FaBars />
-        </button>
-
-        <h2>Hello</h2>
-      </div>
+    <div>
+      <h2>
+        Homepage
+      </h2>
     </div>
   );
 }
